+++ Blog/Blog/Areas/Admin/AdminAreaRegistration.cs
@@ -1,0 +1,24 @@
+﻿using System.Web.Mvc;
+
+namespace Blog.Areas.Admin
+{
+    public class AdminAreaRegistration : AreaRegistration 
+    {
+        public override string AreaName 
+        {
+            get 
+            {
+                return "Admin";
+            }
+        }
+
+        public override void RegisterArea(AreaRegistrationContext context) 
+        {
+            context.MapRoute(
+                "Admin_default",
+                "Admin/{controller}/{action}/{id}",
+                new { action = "Index", id = UrlParameter.Optional }
+            );
+        }
+    }
+}
