+++ Blog/Blog/Areas/Admin/Views/Post/Index.cshtml
@@ -1,0 +1,40 @@
+﻿@model IList<Blog.Models.Post>
+<h2>Lista de Posts</h2>
+@Html.ActionLink("Incluir Post", "NovoPost", null, new { @class = "btn btn-primary" })
+<table class="table table-striped table-hover table-bordered">
+    <caption>Lista dos posts cadastrados no blog.</caption>
+    <thead>
+        <tr>
+            <th>Post</th>
+            <th>Categoria</th>
+            <th>Publicado Em</th>
+            <th>Ações</th>
+        </tr>
+    </thead>
+    <tbody>
+        @foreach (var p in Model)
+        {
+            <tr>
+                <td>
+                    <h4>@Html.ActionLink(p.Titulo, "Visualiza", new { Id = p.Id })</h4>
+                    <p>@p.Resumo</p>
+                </td>
+                <td>@p.Categoria</td>
+                <td>
+                    @if (p.Publicado)
+                    {
+                        @Html.Raw(string.Format("{0:dd/MM/yyyy}", p.DataPublicacao))
+                    }
+                    else
+                    {
+                        @Html.Raw("Não Publicado")
+                    }
+                </td>
+                <td>
+                    @Html.ActionLink("Remover", "RemovePost", new { id = p.Id })
+                    @Html.ActionLink("Publicar", "PublicaPost", new { id = p.Id })
+                </td>
+            </tr>
+        }
+    </tbody>
+</table>
