+++ Blog/Blog/Areas/Admin/Views/Post/Visualiza.cshtml
@@ -1,0 +1,14 @@
+﻿@model Blog.Models.Post
+<h1>Editar Post</h1>
+@using (Html.BeginForm("EditaPost", "Post"))
+{
+    @Html.HiddenFor(p => p.Id)
+    @Html.HiddenFor(p => p.Publicado)
+    @Html.HiddenFor(p => p.DataPublicacao)
+    @Html.EditorFor(p => p)
+    <button type="submit" class="btn btn-primary">Cadastrar</button>
+}
+@section Scripts
+{
+    @Html.Partial("_Scripts")
+}
