+++ Blog/Blog/Areas/Admin/Views/Shared/EditorTemplates/Post.cshtml
@@ -1,0 +1,13 @@
+﻿@model Blog.Models.Post
+
+@Html.LabelFor(p => p.Titulo)
+@Html.TextBoxFor(p => p.Titulo, new { @class = "form-control" })
+@Html.ValidationMessageFor(p => p.Titulo)
+
+@Html.LabelFor(p => p.Resumo)
+@Html.TextAreaFor(p => p.Resumo, new { @class = "form-control", rows = 6 })
+@Html.ValidationMessageFor(p => p.Resumo)
+
+@Html.LabelFor(p => p.Categoria)
+@Html.TextBoxFor(p => p.Categoria, new { @class = "form-control" })
+@Html.ValidationMessageFor(p => p.Categoria)
