+++ Blog/Blog/Areas/Admin/Views/Shared/MenuAdmin.cshtml
@@ -1,0 +1,8 @@
+﻿<aside>
+    <h1>Administração</h1>
+    <div>
+        <ul>
+            <li>@Html.ActionLink("Posts", "Index", "Post")</li>
+        </ul>
+    </div>
+</aside>
