+++ Blog/Blog/Areas/Admin/Views/Shared/_LayoutAdmin.cshtml
@@ -1,0 +1,18 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Administração - Resenha Cultural</title>
+    <link rel="stylesheet" href="~/Content/themes/base/jquery-ui.css" type="text/css" />
+    <link rel="stylesheet" href="~/Content/bootstrap.css" type="text/css" />
+    <link rel="stylesheet" href="~/Content/estilo-admin.css" type="text/css" />
+</head>
+<body>
+    @Html.Partial("MenuAdmin")
+    <main>
+        @RenderBody()
+    </main>
+    @RenderSection("Scripts", false)
+</body>
+</html>
