+++ Blog/Blog/Content/estilo-admin.css
@@ -1,0 +1,34 @@
+body {
+	background-color: #fff;
+}
+main {
+	float: left;
+	width: 75%;
+	padding: 20px;
+}
+aside {
+	float: left;
+	width: 25%;
+	left: 0;
+	padding: 20px;
+	background-color: black;
+	color: white;
+	height: 100vmax;
+}
+
+aside>div{
+	padding-bottom: 40px;
+}
+aside>div>ul{
+	padding-left: 5px;
+}
+main>form+p {
+	padding-top: 10px;
+}
+th {
+	background-color: #ccc;
+	font-weight: bold;
+}
+form button {
+    margin-top: 20px;  
+}
