--- Blog/Blog/Views/Post/Index.cshtml
@@ -1,41 +1,0 @@
-﻿@model IList<Blog.Models.Post>
-
-<div>
-    <h2>Lista de Posts</h2>
-    @Html.ActionLink("Incluir Post", "NovoPost")
-    <table>
-        <thead>
-            <tr>
-                <th>Título</th>
-                <th>Resumo</th>
-                <th>Categoria</th>
-                <th>Publicado?</th>
-                <th>Ações</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var p in Model)
-            {
-                <tr>
-                    <td>@Html.ActionLink(p.Titulo, "Visualiza", new { id = p.Id })</td>
-                    <td>@p.Resumo</td>
-                    <td>@p.Categoria</td>
-                    <td>
-                        @if (p.Publicado)
-                        {
-                            @Html.Raw(string.Format("{0:dd/MM/yyyy}", p.DataPublicacao))
-                        }
-                        else
-                        {
-                            @Html.Raw("Não")
-                        }
-                    </td>
-                    <td>
-                        @Html.ActionLink("Remover", "RemovePost", new { id = p.Id })
-                        @Html.ActionLink("Publicar", "PublicaPost", new { id = p.Id })
-                    </td>
-                </tr>
-            }
-        </tbody>
-    </table>
-</div>
