--- Blog/Blog/Views/Post/NovoPost.cshtml
@@ -1,13 +1,0 @@
-﻿@model Blog.Models.Post
-<div>
-    @using (Html.BeginForm("AdicionaPost", "Post"))
-    {
-        @Html.EditorFor(p => p)
-
-        <button type="submit">Cadastrar</button>
-    }
-</div>
-
-@section Scripts {
-    @Html.Partial("_Scripts")
-}
