--- Blog/Blog/Views/Post/Visualiza.cshtml
@@ -1,18 +1,0 @@
-﻿@model Blog.Models.Post
-<div>
-    @using (Html.BeginForm("EditaPost", "Post"))
-    {
-
-        @Html.HiddenFor(p => p.Id)
-        @Html.HiddenFor(p => p.Publicado)
-        @Html.HiddenFor(p => p.DataPublicacao)
-
-        @Html.EditorFor(p => p)
-
-        <button type="submit">Cadastrar</button>
-    }
-</div>
-
-@section Scripts {
-    @Html.Partial("_Scripts")
-}
