--- Blog/Blog/Views/Shared/EditorTemplates/Post.cshtml
@@ -1,13 +1,0 @@
-﻿@model Blog.Models.Post
-
-@Html.LabelFor(p => p.Titulo)
-@Html.TextBoxFor(p => p.Titulo)
-@Html.ValidationMessageFor(p => p.Titulo)
-
-@Html.LabelFor(p => p.Resumo)
-@Html.TextBoxFor(p => p.Resumo)
-@Html.ValidationMessageFor(p => p.Resumo)
-
-@Html.LabelFor(p => p.Categoria)
-@Html.TextBoxFor(p => p.Categoria)
-@Html.ValidationMessageFor(p => p.Categoria)
