--- Blog/Blog/Views/Shared/_Scripts.cshtml
@@ -1,16 +1,0 @@
-﻿<script src="~/Scripts/jquery-3.3.1.js"></script>
-<script src="~/Scripts/jquery.validate.js"></script>
-<script src="~/Scripts/jquery.validate.unobtrusive.js"></script>
-<script src="~/Scripts/jquery-ui-1.12.1.js"></script>
-<script>
-        $("input[name='Categoria']").autocomplete({
-            source: function (request, response) {
-                $.post(
-                    "@Url.Action("CategoriaAutocomplete")",
-                    { termoDigitado: request.term },
-                    function (data) {
-                        response(data);
-                    });
-            }
-        });
-</script>
